Refactored to follow Adobe API structure
